pyproject.toml
[project]
name = "dockerserverlauncher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3==1.40.57",
    "flask-cors==6.0.1",
    "paramiko==4.0.0",
    "python-dotenv==1.1.1",
]
